+++ server/src/auth/authenticated_user.rs
