+++ server/src/auth/mod.rs
@@ -1,4 +1,6 @@
+mod authenticated_user;
 mod error;
 
 pub mod openid;
+pub use authenticated_user::AuthenticatedUser;
 pub use error::AuthError;
