--- server/src/models/api/mod.rs
@@ -1,5 +1,0 @@
-mod contact;
-mod result;
-
-pub use contact::*;
-pub use result::*;
