--- server/src/models/auth/mod.rs
@@ -1,3 +1,0 @@
-mod authenticated_user;
-
-pub use authenticated_user::AuthenticatedUser;
