+++ server/src/models/location.rs
@@ -1,0 +1,7 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct Location {
+    latitude: f64,
+    longitude: f64,
+}
