+++ server/src/models/mod.rs
@@ -1,3 +1,7 @@
-pub mod api;
-pub mod auth;
-pub mod storage;
+mod location;
+mod timestamp;
+mod user;
+
+pub use location::Location;
+pub use timestamp::Timestamp;
+pub use user::User;
