--- server/src/models/storage/mod.rs
@@ -1,3 +1,0 @@
-mod user;
-
-pub use user::{Location, User};
