+++ server/src/models/timestamp.rs
@@ -1,0 +1,17 @@
+use std::time::{SystemTime, UNIX_EPOCH};
+
+use serde::{Deserialize, Serialize};
+
+#[derive(Serialize, Deserialize)]
+pub struct Timestamp(u64);
+
+impl Timestamp {
+    pub fn now() -> Self {
+        Timestamp(
+            SystemTime::now()
+                .duration_since(UNIX_EPOCH)
+                .expect("SystemTime::now() is prior to the UNIX_EPOCH")
+                .as_secs(),
+        )
+    }
+}
